refactor(examples): extract the Cua tool implementations into cua_sandbox.py
The reference computer handler (GUI primitives, bash with persistent cwd,
read/write/edit/grep/glob with the result formats n2 expects) now lives in
its own module users can copy or adapt; remote_sandbox.py is just the
entrypoint wiring (47 lines).
